pyproject.toml
[tool.poetry]
name = "hw29"
version = "0.1.0"
description = ""
authors = ["nick <7zx.1zx@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
psycopg2-binary = "^2.9.6"
pillow = "^9.5.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
randominfo = "^2.0.2"
faker = "^18.6.0"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
